Pod Setup and Pod Installation
Initiated pod setup to generate pod file. This pod file contains framework specifications which will be used in project
In this project pod file contains frameworks to check network availability using "ReachabilitySwift" and
"SVProgressHUD" is added to show progress indicator to inform activity is in process.
Bridging-Header file is added to the project so as to allow application to use Objective-C classes from "SVProgressHUD" framework in SWIFT classes.
Bridging-Header file name is updated in project target settings against "Objective-C Bridging Header" so as application will know the file's location