Unit Test Cases
Here we have added two test cases for API call. Success test case is designed to give a call to specified API URL and check for its response; if response is valid it leads to successful test else it fails.
Failed test case is designed in a way to check failed network call when API call is given. This two test cases ensures the API call status and let's us inform if we are getting correct response from API or not